.circleci: Prefer using env-file for docker run
'docker run' has the capability to use a file for environment variables,
we should prefer to use that instead of having it be sourced per command
in the docker container.

Also opens the door for cutting down on the total number of commands we
need to echo into a script to then execute as a 'docker exec' command.

Plus side of this approach is that the BASH_ENV is persisted through all
of the steps so there's no need to do any exports / worry about
environment variables not persisting through jobs.

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>

ghstack-source-id: 50549b80178cea0ae4d6f653871fc10cee6e5a06
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43293

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>